% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bldAnnPlt.R
\name{.bldAnnPlt}
\alias{.bldAnnPlt}
\title{Prepare annotation file for legacy data - Plate based}
\usage{
.bldAnnPlt(study.data, vehicle.mapping, posctr.doses, phase)
}
\arguments{
\item{study.data}{Legacy study annotation file from biobanking}

\item{vehicle.mapping}{List of vehicles to treatments mapping}

\item{posctr.doses}{Vector of positive control doses}

\item{phase}{Study phase}
}
\value{
Annotation data table
}
\description{
This function parses plate based legacy data annotation files,
channel, positive control, and assay mapping to build a data.frame with 
compatible format for GladiaTOX loading
}
\details{
Function used only when processing historical data
}
\note{
PMI-specific
}
