% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bldAnnWll.R
\name{.bldAnnWll}
\alias{.bldAnnWll}
\title{Prepare annotation file for legacy data - Well based}
\usage{
.bldAnnWll(study.data, vehicle.mapping, phase)
}
\arguments{
\item{study.data}{Legacy study annotation file from biobanking}

\item{vehicle.mapping}{List of vehicles to treatments mapping}

\item{phase}{Study phase to be registered in GladiaTOC DB}
}
\description{
This function parses well based legacy data annotation files,
channel, positive control, and assay mapping to build a data.frame with
compatible format for GladiaTOX loading
}
\details{
Function used only when processing historical data
}
\note{
PMI-specific
}
