% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAssayTab.R
\name{.buildAssayTab}
\alias{.buildAssayTab}
\title{Prepare assay table mapping info}
\usage{
.buildAssayTab(plate.mtd, chn.map)
}
\arguments{
\item{plate.mtd}{Legacy study annotation file from biobanking}

\item{chn.map}{List of endpoints to thermo channels mapping}
}
\value{
Plate metadata
}
\description{
This function parses plate annotations and create a mapping
between assay endpoints and channels
}
\details{
Function used only when processing historical data
}
\note{
PMI-specific
}
