% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDB.R
\name{.cleanDB}
\alias{.cleanDB}
\title{Remove all data from database for testing purposes}
\usage{
.cleanDB()
}
\value{
None
}
\description{
This function deletes all data from the database and preserves 
the method tables.
}
\note{
PMI-specific
}
