% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteStudy.R
\name{.deleteStudy}
\alias{.deleteStudy}
\title{Completely remove all data for a study}
\usage{
.deleteStudy(asid, db = NULL)
}
\arguments{
\item{asid}{The assay source/study ID}

\item{db}{(optional) the databse to delete from, defaults to the current
database settings}
}
\value{
None
}
\description{
\code{.deleteStudy} completely removes all data for a study from the 
database.
}
\details{
Cannot be undone. Please use carefully. Not exported, as this is
intended for development and should not be used with real data.
}
\note{
PMI-specific
}
