% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoDB.R
\name{.getScanData}
\alias{.getScanData}
\title{Load data from a scan stored on thermoDB}
\usage{
.getScanData(upd, store = "STORE", verbose = TRUE, curlurl = curlurl)
}
\arguments{
\item{upd}{The plate/scan identifier used by ThermoDB}

\item{store}{The ThermoDB store to pull data from}

\item{verbose}{Logical, should curl return the messages to the console?}

\item{curlurl}{webservice URL}
}
\value{
Parsed XML file
}
\description{
This function queries the thermoDB webservices and returns a
data.table with the data for the given scan
}
\note{
PMI-specific
}
