% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line2user.R
\name{.line2user}
\alias{.line2user}
\title{Convert margin line to usr coordinates}
\usage{
.line2user(line, side)
}
\arguments{
\item{line}{Line to convert}

\item{side}{Line side}
}
\value{
none
}
\description{
Convert margin line to usr coordinates
}
