% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoDB.R
\name{.listScans}
\alias{.listScans}
\title{List the scans stored on the thermoDB server}
\usage{
.listScans(store = "STORE", verbose = TRUE, curlurl = curlurl)
}
\arguments{
\item{store}{The ThermoDB store to pull data from}

\item{verbose}{Logical, should curl return the messages to the console?}

\item{curlurl}{webservice URL}
}
\value{
Parsed XML file
}
\description{
This function queries the thermoDB webservices and returns a
data.table with the available scans stored on the server.
}
\note{
PMI-specific
}
