% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoDB.R
\name{.ListsWrapper}
\alias{.ListsWrapper}
\title{List scans and protocols}
\usage{
.ListsWrapper(store = "STORE", verbose = TRUE, curlurl = curlurl)
}
\arguments{
\item{store}{The ThermoDB store to pull data from}

\item{verbose}{Logical, should curl return the messages to the console?}

\item{curlurl}{webservice URL}
}
\value{
List of DB store entries
}
\description{
This function fetched and merges scan list and protocols
}
\note{
PMI-specific
}
