% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBBMap.R
\name{.loadBBMap}
\alias{.loadBBMap}
\title{Register the biobanking fields}
\usage{
.loadBBMap(input)
}
\arguments{
\item{input}{path to JSON file produced by the GUI}
}
\description{
This function parses the output from the GUI after the GladiaTOX
IDs have been loaded into BioBanking and registers the BioBanking IDs into
the 'bb_apid_map' and 'bb_waid_map' fields.
}
\note{
PMI-specific
}
