% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoDB.R
\name{.parseGetScanData}
\alias{.parseGetScanData}
\title{Parse output in .getScanData}
\usage{
.parseGetScanData(xml_file)
}
\arguments{
\item{xml_file}{XML formatted content from Thermo database}
}
\value{
Parsed well information
}
\description{
This function parses the output from the .getScanData function
in the ThermoDB webservices.
}
\note{
PMI-specific
}
