% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoDB.R
\name{.parseListScans}
\alias{.parseListScans}
\title{Parse output in .listScans}
\usage{
.parseListScans(xml_file)
}
\arguments{
\item{xml_file}{XML formatted content from Thermo database}
}
\value{
DB store entry table
}
\description{
This function parses the output from the .listScans function
in the ThermoDB webservices.
}
\note{
PMI-specific
}
