% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermoDB.R
\name{.parseWellScan}
\alias{.parseWellScan}
\title{helper function for .parseGetScanData}
\usage{
.parseWellScan(l)
}
\arguments{
\item{l}{Content of a single well}
}
\description{
This function parses the individual wells within the
.parseGetScanData function
}
\note{
PMI-specific
}
