% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInput.R
\name{.readInput}
\alias{.readInput}
\title{Parse GUI JSON output}
\usage{
.readInput(x)
}
\arguments{
\item{x}{path to JSON file produced by the GUI}
}
\description{
Parse GUI JSON output
}
\details{
Will also accept a data.frame and convert it to data.table
}
\note{
PMI-specific
}
