% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResultForToxpiGUI.R
\name{exportResultForToxpiGUI}
\alias{exportResultForToxpiGUI}
\title{Create the result table for the asi in input}
\usage{
exportResultForToxpiGUI(asid, tp, outfile, stat)
}
\arguments{
\item{asid}{Assay source id}

\item{tp}{Time point}

\item{outfile}{Path to the output file}

\item{stat}{Character vector of statistic to export}
}
\value{
None
}
\description{
This function export results
}
\details{
This funtion is useful to export results in a table format
}
\examples{
## Export MEC (or AC50) values to be visualized in ToxPiGUI
conf_store <- gtoxConfList()
gtoxConfDefault()

out <- "export_for_toxpiGUI.csv"
exportResultForToxpiGUI(asid=1L, tp="4h", outfile=out, stat=quote(modl_acc))

## Reset configuration
options(conf_store)

}
