% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glCheckInput.R
\name{glCheckInput}
\alias{glCheckInput}
\title{Check validity of input file}
\usage{
glCheckInput(file)
}
\arguments{
\item{file}{file URL}
}
\value{
List of error messages in JSON format
}
\description{
This function check the structure and content of an input file.
}
\details{
This function is useful to check the structure and content of an input file
from the GladiaTOX GUI
}
