% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glLoadInput.R
\name{glLoadInput}
\alias{glLoadInput}
\title{Check validity of input file}
\usage{
glLoadInput(file = NULL, studyname = "samplestudy",
  phasename = "samplephase", tab = NULL)
}
\arguments{
\item{file}{file URL}

\item{studyname}{Name of the study}

\item{phasename}{Name of the phase}

\item{tab}{input table is file URL is not provided}
}
\value{
List of error messages in JSON format
}
\description{
This function check the structure and content of an input file.
}
\details{
This function is useful to load an input file in the GladiaTOX GUI
}
