% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glPlotPieLogo.R
\name{glPlotPieLogo}
\alias{glPlotPieLogo}
\title{plot package logo}
\usage{
glPlotPieLogo()
}
\value{
None
}
\description{
This function plots the GladiaTOX logo.
}
\details{
This funtion is only used to plot the package logo.
}
\examples{
glPlotPieLogo()

}
