% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxListFlds.R
\name{gtoxListFlds}
\alias{gtoxListFlds}
\title{Load the field names for a table}
\usage{
gtoxListFlds(tbl, db = getOption("TCPL_DB"))
}
\arguments{
\item{tbl}{Character of length 1, the gtox database table}

\item{db}{Character of length 1, the gtox database}
}
\value{
A string of field names for the given table.
}
\description{
\code{gtoxListFlds} loads the column names for the given table and database.
}
\details{
This function can be particularly useful in defining the 'fld' param in the
gtoxLoad- functions.
}
\examples{
## Gives the fields in the mc1 table
gtoxListFlds("mc1")

}

