% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxLoadApid.R
\name{gtoxLoadApid}
\alias{gtoxLoadApid}
\title{Load assay plate information}
\usage{
gtoxLoadApid(fld = NULL, val = NULL)
}
\arguments{
\item{fld}{Character, the field(s) to query on}

\item{val}{List, vectors of values for each field to query on. Must be in
the same order as 'fld'.}
}
\value{
A data.table with the assay plate information for the given
parameters
}
\description{
\code{gtoxLoadApid} queries the gtox database and returns the assay plate
information for the given field and values.
}
\examples{
## Store the current config settings, so they can be reloaded at the end 
## of the examples
conf_store <- gtoxConfList()
gtoxConfDefault()

## Prepare for analysis before QC + process data
gtoxLoadApid()

## Reset configuration
options(conf_store)

}

