% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxLoadUnit.R
\name{gtoxLoadUnit}
\alias{gtoxLoadUnit}
\title{Load response units for assay endpoints}
\usage{
gtoxLoadUnit(aeid)
}
\arguments{
\item{aeid}{Integer, assay endpoint ids}
}
\value{
A data.table containing level 3 correction methods for the given
aeids.
}
\description{
\code{gtoxLoadUnit} queries the gtox databases and returns a data.table
with the response units for the given assay endpoint ids (aeid).
}
\seealso{
\code{\link{gtoxQuery}}, \code{\link{data.table}}
}
