% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxLoadVehicle.R
\name{gtoxLoadVehicle}
\alias{gtoxLoadVehicle}
\title{Load vehicle information}
\usage{
gtoxLoadVehicle(field = NULL, val = NULL)
}
\arguments{
\item{field}{Character of length 1, the field to query on}

\item{val}{Vector of values to subset on}
}
\value{
A data.table with the list of vehicles and vehicles ids.
}
\description{
\code{gtoxLoadVehicle} queries the gtox database and returns the vehicle
information for the given field and values.
}
\examples{
## Store the current config settings, so they can be reloaded at the end 
## of the examples
conf_store <- tcplConfList()
tcplConfDefault()

## Prepare for analysis before QC + process data
gtoxLoadVehicle()

## Reset configuration
options(conf_store)

}
