% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxLoadVmad.R
\name{gtoxLoadVmad}
\alias{gtoxLoadVmad}
\title{Load cutoff values for assay endpoints}
\usage{
gtoxLoadVmad(aeid = NULL)
}
\arguments{
\item{aeid}{Integer, assay endpoint ids}
}
\value{
A data.table containing cutoff values for the given aeids.
}
\description{
\code{gtoxLoadVmad} queries the gtox databases and returns a data.table
with the cutoff values for the given assay endpoint ids (aeid).
}
\examples{
## Store the current config settings, so they can be reloaded at the end 
## of the examples
conf_store <- tcplConfList()
tcplConfDefault()

## Prepare for analysis before QC + process data
gtoxLoadVmad()

## Reset configuration
options(conf_store)

}
