% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxPlotErrBar.R
\name{gtoxPlotErrBar}
\alias{gtoxPlotErrBar}
\title{Create error bar plots}
\usage{
gtoxPlotErrBar(c1, c2, aeid, ngrp = NULL)
}
\arguments{
\item{c1}{Integer of length 1, the chid value for the first chemical}

\item{c2}{Integer of length 1, the chid value for the first chemical}

\item{aeid}{Integer, the aeid value(s) to plot}

\item{ngrp}{Integer, the number of "slots" to draw; overridden if the
number of aeid values is greater than 'ngrp'}
}
\value{
None
}
\description{
\code{gtoxPlotErrBar} creates the error bar plots.
}
\examples{
## Make plot

tcplConfDefault()
gtoxPlotErrBar(1,3,2)

}
