% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxPlotPie.R
\name{gtoxPlotPie}
\alias{gtoxPlotPie}
\title{Create piechart plots}
\usage{
gtoxPlotPie(chid, mrks, aeid, col = NULL, lbl = NULL)
}
\arguments{
\item{chid}{Integer of length 1, the chid value}

\item{mrks}{Numeric, the values for concentration label rings}

\item{aeid}{Integer, the aeid values to plot}

\item{col}{Vector of colors}

\item{lbl}{Vector with pie labels (optional)}
}
\value{
None
}
\description{
\code{gtoxPlotPie} creates the piechart plots.
}
\examples{

## Plot pie for chemical IDs 1 to 5 and multiple endpoints
gtoxPlotPie(chid=1:5, mrks=10^c(1:6), aeid=c(2:10))

}
