% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxPlotPieLgnd.R
\name{gtoxPlotPieLgnd}
\alias{gtoxPlotPieLgnd}
\title{Create piechart plot legend}
\usage{
gtoxPlotPieLgnd(aenm, ncol = 2, col = NULL, fit.labels = TRUE)
}
\arguments{
\item{aenm}{Character, the assay endpoint names}

\item{ncol}{Interger, the number of columns for the legend}

\item{col}{Vector of colors}

\item{fit.labels}{Boolean, if TRUE, scale the text to fit}
}
\value{
None
}
\description{
\code{gtoxPlotPieLgnd} creates the piechart plots.
}

