% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw.R
\name{lw}
\alias{lw}
\title{Abbreviation for \code{length(which(x))}}
\usage{
lw(x)
}
\arguments{
\item{x}{A logical}
}
\value{
The length of the \code{TRUE} values in \code{x}
}
\description{
\code{lw} takes a logical vector, \code{x}, and returns 
\code{length(which(x))}.
}
\examples{
lw(c(TRUE, FALSE, TRUE))

}
\seealso{
\code{\link{length}}, \code{\link{which}}

Other gtox abbreviations: \code{\link{is.odd}},
\code{\link{lu}}, \code{\link{sink_reset}}
}
\concept{gtox abbreviations}
