% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_reset.R
\name{sink_reset}
\alias{sink_reset}
\title{Reset all sinks}
\usage{
sink_reset()
}
\value{
None
}
\description{
\code{sink_reset} resets all sinks and returns all output to the console.
}
\details{
\code{sink_reset} identifies all sinks with \code{sink.number} then returns
all output and messages back to the console.
}
\seealso{
\code{\link{sink}}, \code{\link{sink.number}}

Other gtox abbreviations: \code{\link{is.odd}},
  \code{\link{lu}}, \code{\link{lw}}
}
\concept{gtox abbreviations}
