% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplConf.R
\name{tcplConf}
\alias{tcplConf}
\title{Functions for configuring the gtox package}
\usage{
tcplConf(drvr = NULL, user = NULL, pass = NULL, host = NULL,
  db = NULL, int = NULL)
}
\arguments{
\item{drvr}{Character of length 1, which database driver to use}

\item{user}{Character of length 1, the database server username}

\item{pass}{Character of length 1, the database server password}

\item{host}{Character of length 1, the database server}

\item{db}{Character of length 1, the name of the tcpl database}

\item{int}{Logical}
}
\value{
none
}
\description{
This function is used to set default configuration of the gtox sqlite DB.
}
\examples{

tcplConf()

}
