% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplConfDefault.R
\name{tcplConfDefault}
\alias{tcplConfDefault}
\title{Functions for configuring the gtox package}
\usage{
tcplConfDefault()
}
\value{
none
}
\description{
This function is used to set default configuration of the gtox sqlite DB.
}
