% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplConfReset.R
\name{tcplConfReset}
\alias{tcplConfReset}
\title{Functions for configuring the gtox package}
\usage{
tcplConfReset()
}
\value{
none
}
\description{
This function is used to reset gtox sqlite DB configuration.
}
