% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplConfSave.R
\name{tcplConfSave}
\alias{tcplConfSave}
\title{Functions for configuring the gtox package}
\usage{
tcplConfSave()
}
\value{
none
}
\description{
This function is used to set default configuration of the gtox sqlite DB.
}
