% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplLoadVmad.R
\name{tcplLoadVmad}
\alias{tcplLoadVmad}
\title{Load cutoff values for assay endpoints}
\usage{
tcplLoadVmad(aeid = NULL)
}
\arguments{
\item{aeid}{Integer, assay endpoint ids}
}
\value{
A data.table containing cutoff values for the given aeids.
}
\description{
\code{tcplLoadVmad} queries the tcpl databases and returns a data.table
with the cutoff values for the given assay endpoint ids (aeid).
}
\examples{
## Store the current config settings, so they can be reloaded at the end 
## of the examples
conf_store <- tcplConfList()
tcplConfDefault()

## Prepare for analysis before QC + process data
tcplLoadVmad()

## Reset configuration
options(conf_store)

}
