% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotFitc.R
\name{tcplPlotFitc}
\alias{tcplPlotFitc}
\title{Plot the fit category tree}
\usage{
tcplPlotFitc(fitc = NULL, main = NULL, fitc_sub = NULL)
}
\arguments{
\item{fitc}{Integer, the fit categories}

\item{main}{Character of length 1, the title (optional)}

\item{fitc_sub, }{Integer, a subset of fit categories to plot}
}
\value{
None
}
\description{
\code{tcplPlotFitc} makes a plot showing the level 5 fit categories.
}
\note{
Suggested device size (inches): width = 10, height = 7.5, pointsize = 9
}
\examples{
## Store the current config settings, so they can be reloaded at the end
## of the examples
conf_store <- tcplConfList()
tcplConfDefault()

## Display the fit category tree.
tcplPlotFitc()

}
