% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotPie.R
\name{tcplPlotPie}
\alias{tcplPlotPie}
\title{Create piechart plots}
\usage{
tcplPlotPie(chid, mrks, aeid, col = NULL, lbl = NULL)
}
\arguments{
\item{chid}{Integer of length 1, the chid value}

\item{mrks}{Numeric, the values for concentration label rings}

\item{aeid}{Integer, the aeid values to plot}

\item{col}{Vector of colors}

\item{lbl}{Vector with pie labels (optional)}
}
\value{
None
}
\description{
\code{tcplPlotPie} creates the piechart plots.
}
