% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotPieLgnd.R
\name{tcplPlotPieLgnd}
\alias{tcplPlotPieLgnd}
\title{Create piechart plot legend}
\usage{
tcplPlotPieLgnd(aenm, ncol = 2, col = NULL, fit.labels = TRUE)
}
\arguments{
\item{aenm}{Character, the assay endpoint names}

\item{ncol}{Interger, the number of columns for the legend}

\item{col}{Vector of colors}

\item{fit.labels}{Boolean, if TRUE, scale the text to fit}
}
\value{
None
}
\description{
\code{tcplPlotPieLgnd} creates the piechart plots.
}
