% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotWin.R
\name{tcplPlotWin}
\alias{tcplPlotWin}
\title{Create winning curve plots}
\usage{
tcplPlotWin(chid, aeid, bline = "bmad", collapse = TRUE)
}
\arguments{
\item{chid}{Integer of length 1, the chid value}

\item{aeid}{Integer, the aeid values to plot}

\item{bline}{Character of length 1, the value used for drawing the baseline
noise}

\item{collapse}{Logical, collapse the data by spid when true}
}
\value{
None
}
\description{
\code{tcplPlotWin} creates the piechart plots.
}
\details{
When 'collapse' is TRUE the plotted points will be the mean of the values
based on spid.

Any values for 'bline' other than 'coff' will use 3*bmad.
}
\examples{

chid <- tcplLoadChem(field="chnm", val="acrylamide", include.spid=FALSE)$chid
print(chid)
aeid <- tcplLoadAeid(fld=c("asid","aenm"), 
          val=list(1L, "GSH content_GSH content_4h_dn"), add.fld="asid")$aeid
print(aeid)
tcplPlotWin(chid = chid, aeid = aeid, bline="bmad", collapse=TRUE)

}

