% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplQuery.R
\name{tcplQuery}
\alias{tcplQuery}
\title{Wrapper for sending queries and fetching results}
\usage{
tcplQuery(query, db = getOption("TCPL_DB"),
  drvr = getOption("TCPL_DRVR"), tbl = NULL)
}
\arguments{
\item{query}{Character of length 1, the query string}

\item{db}{Character of length 1, the name of the tcpl database}

\item{drvr}{Character of length 1, which database driver to use}

\item{tbl}{Added for compatibility with tcpl last version}
}
\value{
none
}
\description{
These functions send a query to the given database
}
\examples{
qformat <- "SELECT * FROM assay_source"
tcplQuery(qformat)

}
