% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glPlotPosCtrl.R
\name{glPlotPosCtrl}
\alias{glPlotPosCtrl}
\title{Box plot for positive control check}
\usage{
glPlotPosCtrl(asid)
}
\arguments{
\item{asid}{Assay source id}
}
\value{
A list of ggplot objects, one per assay X timepoint.
}
\description{
This function plots positive controls as well as vehicle and
treatments normalized values
}
\details{
This funtion is useful to select plates to mask
}
\examples{
## Store the current config settings, so they can be reloaded at the end
## of the examples
conf_store <- gtoxConfList()
gtoxConfDefault()

## Create boxplot for all endpoints and chemicals tested. Useful to save
## plots in a pdf file.
pp <- glPlotPosCtrl(asid = 1L)
pp[[1]]

## Reset configuration
options(conf_store)

}
