% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoxPlotPie.R
\name{gtoxPlotPie}
\alias{gtoxPlotPie}
\title{Create piechart plots}
\usage{
gtoxPlotPie(chid, mrks, aeid, col = NULL, lbl = NULL)
}
\arguments{
\item{chid}{Integer of length 1, the chid value}

\item{mrks}{Numeric, the values for concentration label rings}

\item{aeid}{Integer, the aeid values to plot}

\item{col}{Vector of colors}

\item{lbl}{Vector with pie labels (optional)}
}
\value{
None
}
\description{
\code{gtoxPlotPie} creates the piechart plots.
}
\examples{

## Create a pie plot for report
tcplConfDefault()
gtoxPlotPie(3, 2, 1:5)

}
