% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDatForDB.R
\name{prepareDatForDB}
\alias{prepareDatForDB}
\title{Assign default processing methods}
\usage{
prepareDatForDB(asid, dat)
}
\arguments{
\item{asid}{Integer, the asid value(s) to assign the default methods to}

\item{dat}{Data.table containing metadata and data to load in DB}
}
\value{
Data table with data and metadata to store in database
}
\description{
This function is a wrapper to ease the creation of the 
dataframe containing data and metadata to be loaded in the database
}
\details{
This function formats a dat table to be loaded in DB
}
\examples{

\dontrun{
dat <- prepareDatForDB(asid, dat)
}

}

